+++ src/bezel/status.ts
@@ -80,10 +80,9 @@ export abstract class RunnableComponent<T extends ComponentConfiguration>
       return;
     }
     if (this._status === Status.DISABLED) {
-      console.log(`${this.name} skip status change (currently disabled) => ${status}`);
       return;
     }
-    console.log(`${this.name} status change: ${this._status} => ${status}`);
+    // console.log(`${this.name} status change: ${this._status} => ${status}`);
     this._status = status;
     this._onDidChangeStatus.fire(status);
   }
@@ -154,6 +153,7 @@ export interface LaunchArgs {
   command: string[];
   showSuccessfulLaunchTerminal: boolean;
   showFailedLaunchTerminal: boolean;
+  cwd?: string;
 }
 
 /**
@@ -230,7 +230,7 @@ export abstract class LaunchableComponent<
   /**
    * getLaunchArgs should return the command line arguments.
    */
-  abstract getLaunchArgs(): Promise<LaunchArgs>;
+  abstract getLaunchArgs(): Promise<LaunchArgs | undefined>;
 
   setDisabled(b: boolean) {
     super.setDisabled(b);
@@ -326,6 +326,9 @@ export abstract class LaunchableComponent<
 
   async handleCommandLaunch(extraArgs: string[] = []): Promise<void> {
     const launchArgs = await this.getLaunchArgs();
+    if (!launchArgs) {
+      return;
+    }
 
     if (!this.terminal) {
       console.log(`Launching terminal "${this.terminalName}"...`);
@@ -357,7 +360,7 @@ export abstract class LaunchableComponent<
           await this.launchInternal();
           clearTimeout(timeout);
           this.handleLaunchSuccess(launchArgs, this.terminal!);
-          resolve()
+          resolve();
         } catch (err) {
           if (--iteration <= 0) {
             clearTimeout(timeout);