+++ src/extension.ts
@@ -6,20 +6,27 @@ import { BazelStackVSCodeAPI } from 'bazel-stack-vscode-api';
 import { BzlFeature } from './bezel/feature';
 import { Telemetry, CommandName, openExtensionSetting } from './constants';
 import { Container } from './container';
+import { ConfigurationContext } from './common';
 
 const api = new API();
 
 export function activate(ctx: vscode.ExtensionContext): BazelStackVSCodeAPI {
   try {
-    Container.initialize(ctx);
+    const configCtx = new ConfigurationContext(
+      ctx.extensionUri,
+      ctx.globalStorageUri,
+      ctx.workspaceState,
+    );
+    Container.initialize(configCtx, ctx.subscriptions);
 
     ctx.subscriptions.push(
       vscode.commands.registerCommand(CommandName.OpenSetting, openExtensionSetting)
     );
 
+
     ctx.subscriptions.push(new BazelDocFeature());
-    ctx.subscriptions.push(new BazelrcFeature());
-    ctx.subscriptions.push(new BzlFeature(api, ctx));
+    ctx.subscriptions.push(new BazelrcFeature(configCtx));
+    ctx.subscriptions.push(new BzlFeature(api, ctx, configCtx));
 
     Container.telemetry.sendTelemetryEvent(Telemetry.ExtensionActivate);