+++ src/test/integration/feature.bazeldoc.test.ts
@@ -90,15 +90,15 @@ suite(BazelDocFeatureName, function () {
     const md = hover?.contents[0] as vscode.MarkdownString;
     expect(md.value).to.be.equal(
       '**[http_archive](https://example.com/be/http.html#http_archive)** is a member of _http_archive_' +
-        '\n\n' +
-        '[http_archive](https://example.com/be/http.html#http_archive), [http_jar](https://example.com/be/http.html#http_jar)'
+      '\n\n' +
+      '[http_archive](https://example.com/be/http.html#http_archive), [http_jar](https://example.com/be/http.html#http_jar)'
     );
   });
 
   // skip this test as it throws an error
-  // test.skip('not hover if feature disabled)', async () => {
-  // 	await featureConfig.update('enabled', false);
-  // 	const hover = await hoverProvider.provideHover(document, new vscode.Position(1, 5), cancellationTokenSource.token);
-  // 	expect(hover).to.be.undefined;
-  // });
+  test.skip('not hover if feature disabled)', async () => {
+    await featureConfig.update('enabled', false);
+    const hover = await hoverProvider.provideHover(document, new vscode.Position(1, 5), cancellationTokenSource.token);
+    expect(hover).to.be.undefined;
+  });
 });