+++ src/test/integration/feature.bazelrc.test.ts
@@ -7,6 +7,10 @@ import { expect } from 'chai';
 import { after, before, describe, it } from 'mocha';
 import { BazelrcFeatureName } from '../../bazelrc/feature';
 import { BazelFlagSupport } from '../../bazelrc/flags';
+import path = require('path');
+import { ConfigurationContext } from '../../common';
+import { FakeMemento } from '../memento';
+import { Container } from '../../container';
 
 tmp.setGracefulCleanup();
 
@@ -26,21 +30,22 @@ type flagCompletionTest = {
   numItems?: number; // expected number of completion items
 };
 
-type codelensTest = {
-  d: string; // test description
-  input: string; // content of file
-  numItems?: number; // expected number of codelens items
-  command?: vscode.Command; // expected command
-  range?: vscode.Range; // the expected range, if we care
-};
-
 describe(BazelrcFeatureName, function () {
   let support: BazelFlagSupport;
   const cancellationTokenSource = new vscode.CancellationTokenSource();
 
+  const extensionPath = path.join(__dirname, '..', '..', '..');
+  const extensionUri = vscode.Uri.file(extensionPath);
+  const configCtx = new ConfigurationContext(
+    extensionUri,
+    vscode.Uri.file(''), // globalStorageUri should not be needed for test
+    new FakeMemento(),
+  );
+  Container.initialize(configCtx, []);
+
   before(async () => {
     const emitter = new vscode.EventEmitter<void>();
-    support = new BazelFlagSupport(emitter.event);
+    support = new BazelFlagSupport(configCtx, emitter.event);
     emitter.fire();
   });